+++ frontend/src/landing/HomeSections/Benefits.tsx
@@ -12,13 +12,13 @@ import useTranslation from 'next-translate/useTranslation';
 import lottie from 'lottie-web';
 
 import { Lottie } from '@/types/animations';
-import { Column, Description, Row, Wrapper } from '@/ui/pages';
+import { Column, Description, Row, Wrapper } from '@/landing/ui/styles';
 
 interface Props {
   item: Lottie;
 }
 
-const importAnimatiom = (path: string) => require(`@/ui/animations/home/${path}`);
+const importAnimatiom = (path: string) => require(`@/landing/ui/animations/home/${path}`);
 
 const Benefits = ({ item }: Props): JSX.Element => {
   const { t } = useTranslation('home');
