+++ frontend/src/landing/ui/animations/home/banner-morph.json
