+++ frontend/src/landing/ui/animations/home/billboard-desktop.json
