+++ frontend/src/landing/ui/animations/home/billboard-mobile.json
