+++ frontend/src/landing/ui/animations/home/keybenefit2.json
